feature/lab 6 basic 1

+++ lab6/lib/include/string.h
@@ -1,8 +1,10 @@
+#include "stdint.h"
+
 int strlen(char *s);
 int atoi(char* str);
 void append_str(char *s, char ch);
 void pop_str(char *s);
 int strcmp(const char *X, const char *Y);
-int myHex2Int(char* str);
+uint64_t myHex2Int(char* str);
 char *strtok(char *str, char *delimiter);
 int spilt_strings(char** str_arr, char* str, char* deli);
